Implemented MMAS properly, updated optimal values (to fit standard ACO) added a timer class, updated visual stuff to update labels

+++ src/main/java/group/ea/controllers/Schedule.java
@@ -179,7 +179,7 @@ public class Schedule {
             case "Ant System":
                 System.out.println("ANT");
                 this.algorithm = new ACO(this.searchSpace, this.problem);
-                this.algorithm.setValues(Double.parseDouble(optionalValues[0]), Double.parseDouble(optionalValues[1]), Integer.parseInt(optionalValues[2]));
+                this.algorithm.setValues(Integer.parseInt(optionalValues[0]), Double.parseDouble(optionalValues[1]), Double.parseDouble(optionalValues[2]));
                 System.out.println(optionalValues[0]);
                 break;
             default:
