+++ src/main/java/group/ea/structure/algorithm/Algorithm.java
@@ -111,7 +111,7 @@ public abstract class Algorithm {
         this.listener = controller;
         //System.out.println("Listener set" + listener + " Controller sent= "+ controller);
     }
-    public void setValues(double a, double b, double r){
+    public  void setValues(int a, double b, double r){
 
     }
 }
